Refactro dns crate
